Move GObject, GError, GInetSocketAddress, GSocket and Closure into GLib sub-package.
