+++ src/org/freedesktop/gstreamer/glib/GSocketAddress.java
@@ -1,7 +1,5 @@
 package org.freedesktop.gstreamer.glib;
 
-import org.freedesktop.gstreamer.GObject;
-
 public class GSocketAddress extends GObject{
 
 	public static final String GTYPE_NAME = "GSocketAddress";
