+++ test/org/freedesktop/gstreamer/PipelineTest.java
@@ -20,6 +20,8 @@
 
 package org.freedesktop.gstreamer;
 
+import org.freedesktop.gstreamer.glib.GError;
+import org.freedesktop.gstreamer.glib.GObject;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
